GenomicCoordinates

Documentation for GenomicCoordinates.

GenomicCoordinates.GenomicInterval

GenomicCoordinates.GenomicInterval

GenomicCoordinates.GenomicPosition

GenomicCoordinates.chr2int

GenomicCoordinates.compare_for_inclusion_1in2

GenomicCoordinates.compare_for_inclusion_2in1

GenomicCoordinates.compare_for_overlap

GenomicCoordinates.find_intersections

GenomicCoordinates.segment_length

GenomicCoordinates.GenomicInterval — Type

GenomicInterval{C,P}

A genomic interval.

GenomicCoordinates.GenomicInterval — Method

GenomicInterval(chr, first::P, last::P)

Create a GenomicInterval with chromosome chr, start position first, and end position last.

GenomicCoordinates.GenomicPosition — Type

GenomicPosition{C,P}

A genomic position.

GenomicCoordinates.chr2int — Method

chr2int(chr)::Int

Convert a chromosome name to an integer.

GenomicCoordinates.compare_for_inclusion_1in2 — Method

compare_for_inclusion_1in2(i1::Interval, i2::Interval)

Compare two intervals for inclusion. Returns -2 if i1 is before i2, 2 if i1 is after i2, -1 if i1 partially overlaps i2 at the beginning, 1 if i1 partially overlaps i2 at the end, and 0 if i1 is inside i2.

GenomicCoordinates.compare_for_inclusion_2in1 — Method

compare_for_inclusion_2in1(i1::Interval, i2::Interval)

Compare two intervals for inclusion. Returns -2 if i1 is before i2, 2 if i1 is after i2, -1 if i2 partially overlaps i1 at the beginning, 1 if i2 partially overlaps i1 at the end, and 0 if i2 is inside i1.

GenomicCoordinates.compare_for_overlap — Method

compare_for_overlap(i1::Interval, i2::Interval)

Compare two intervals for overlap. Returns -1 if i1 is before i2, 1 if i1 is after i2, and 0 if they overlap.

GenomicCoordinates.find_intersections — Function

find_intersections(x, y, compare=compare_for_overlap)

Find intersections between two arrays of intervals. The compare function is used to compare two intervals. The default is compare_for_overlap.

Returns an array of arrays, where the i-th element contains the indices of intervals in y that intersect with the i-th interval in x.

The intervals in x and y do not need to be sorted. However the function will sort them internally, so for repeated calls it is more efficient to sort them before calling this function.

GenomicCoordinates.segment_length — Method

segment_length(a::GenomicInterval{C,P})

Return the length of the genomic segment a.